+++ packages/backend/src/services/ProjectService/ProjectService.ts
@@ -1,18 +1,29 @@
 import { subject } from '@casl/ability';
 import {
+    addDashboardFiltersToMetricQuery,
     AlreadyExistsError,
     AlreadyProcessingError,
     AndFilterGroup,
     AnyType,
     ApiChartAndResults,
+    type ApiCreateProjectResults,
+    type ApiExecuteAsyncQueryResults,
+    type ApiGetAsyncQueryResults,
     ApiQueryResults,
     ApiSqlQueryResults,
+    assertUnreachable,
     CacheMetadata,
+    type CalculateSubtotalsFromQuery,
     CalculateTotalFromQuery,
     ChartSourceType,
     ChartSummary,
     CompiledDimension,
     ContentType,
+    convertCustomMetricToDbt,
+    countCustomDimensionsInMetricQuery,
+    countTotalFilterRules,
+    type CreateDatabricksCredentials,
+    createDimensionWithGranularity,
     CreateJob,
     CreateProject,
     CreateProjectMember,
@@ -20,7 +31,6 @@ import {
     CreateVirtualViewPayload,
     CreateWarehouseCredentials,
     CustomFormatType,
-    DEFAULT_RESULTS_PAGE_SIZE,
     DashboardAvailableFilters,
     DashboardBasicDetails,
     DashboardFilters,
@@ -28,77 +38,24 @@ import {
     DbtExposure,
     DbtExposureType,
     DbtProjectType,
+    deepEqual,
+    DEFAULT_RESULTS_PAGE_SIZE,
     DefaultSupportedDbtVersion,
     DimensionType,
     DownloadFileType,
+    type ExecuteAsyncDashboardChartRequestParams,
+    type ExecuteAsyncMetricQueryRequestParams,
+    type ExecuteAsyncQueryRequestParams,
+    type ExecuteAsyncSavedChartRequestParams,
     Explore,
     ExploreError,
     ExploreType,
+    FilterableDimension,
     FilterGroupItem,
     FilterOperator,
-    FilterableDimension,
-    ForbiddenError,
-    IntrinsicUserAttributes,
-    ItemsMap,
-    Job,
-    JobStatusType,
-    JobStepType,
-    JobType,
-    LightdashError,
-    MetricQuery,
-    MissingWarehouseCredentialsError,
-    MostPopularAndRecentlyUpdated,
-    NotExistsError,
-    NotFoundError,
-    ParameterError,
-    PivotChartData,
-    PivotValuesColumn,
-    Project,
-    ProjectCatalog,
-    ProjectGroupAccess,
-    ProjectMemberProfile,
-    ProjectMemberRole,
-    ProjectType,
-    QueryExecutionContext,
-    ReplaceCustomFields,
-    ReplaceCustomFieldsPayload,
-    ReplaceableCustomFields,
-    RequestMethod,
-    ResultRow,
-    SavedChartDAO,
-    SavedChartsInfoForDashboardAvailableFilters,
-    SessionUser,
-    SortByDirection,
-    SortField,
-    SpaceQuery,
-    SpaceSummary,
-    SqlRunnerPayload,
-    SqlRunnerPivotQueryPayload,
-    SummaryExplore,
-    TableSelectionType,
-    TablesConfiguration,
-    UnexpectedServerError,
-    UpdateMetadata,
-    UpdateProject,
-    UpdateProjectMember,
-    UpdateVirtualViewPayload,
-    UserAttributeValueMap,
-    UserWarehouseCredentials,
-    VizColumn,
-    WarehouseClient,
-    WarehouseCredentials,
-    WarehouseTableSchema,
-    WarehouseTablesCatalog,
-    WarehouseTypes,
-    addDashboardFiltersToMetricQuery,
-    assertUnreachable,
-    convertCustomMetricToDbt,
-    countCustomDimensionsInMetricQuery,
-    countTotalFilterRules,
-    createDimensionWithGranularity,
-    deepEqual,
     findFieldByIdInExplore,
     findReplaceableCustomMetrics,
+    ForbiddenError,
     formatRawRows,
     formatRow,
     formatRows,
@@ -115,28 +72,73 @@ import {
     getSubtotalKey,
     getTimezoneLabel,
     hasIntersection,
+    IntrinsicUserAttributes,
     isCustomSqlDimension,
     isDateItem,
     isDimension,
     isExploreError,
-    isFilterRule,
     isFilterableDimension,
+    isFilterRule,
     isNotNull,
     isUserWithOrg,
+    ItemsMap,
+    Job,
+    JobStatusType,
+    JobStepType,
+    JobType,
+    LightdashError,
     maybeReplaceFieldsInChartVersion,
+    MetricQuery,
+    MissingWarehouseCredentialsError,
+    MostPopularAndRecentlyUpdated,
+    NotExistsError,
+    NotFoundError,
+    ParameterError,
+    PivotChartData,
+    PivotValuesColumn,
+    Project,
+    ProjectCatalog,
+    ProjectGroupAccess,
+    ProjectMemberProfile,
+    ProjectMemberRole,
+    ProjectType,
+    QueryExecutionContext,
+    QueryHistoryStatus,
+    ReplaceableCustomFields,
+    ReplaceCustomFields,
+    ReplaceCustomFieldsPayload,
     replaceDimensionInExplore,
-    snakeCaseName,
-    type ApiCreateProjectResults,
-    type ApiPaginatedQueryResults,
-    type CalculateSubtotalsFromQuery,
-    type CreateDatabricksCredentials,
-    type PaginatedDashboardChartRequestParams,
-    type PaginatedMetricQueryRequestParams,
-    type PaginatedQueryRequestParams,
-    type PaginatedSavedChartRequestParams,
+    RequestMethod,
+    ResultRow,
     type RunQueryTags,
+    SavedChartDAO,
+    SavedChartsInfoForDashboardAvailableFilters,
     type SemanticLayerConnectionUpdate,
+    SessionUser,
+    snakeCaseName,
+    SortByDirection,
+    SortField,
+    SpaceQuery,
+    SpaceSummary,
+    SqlRunnerPayload,
+    SqlRunnerPivotQueryPayload,
+    SummaryExplore,
+    TablesConfiguration,
+    TableSelectionType,
     type Tag,
+    UnexpectedServerError,
+    UpdateMetadata,
+    UpdateProject,
+    UpdateProjectMember,
+    UpdateVirtualViewPayload,
+    UserAttributeValueMap,
+    UserWarehouseCredentials,
+    VizColumn,
+    WarehouseClient,
+    WarehouseCredentials,
+    WarehouseTablesCatalog,
+    WarehouseTableSchema,
+    WarehouseTypes,
 } from '@lightdash/common';
 import { SshTunnel } from '@lightdash/warehouses';
 import * as Sentry from '@sentry/node';
@@ -177,9 +179,9 @@ import { UserWarehouseCredentialsModel } from '../../models/UserWarehouseCredent
 import { WarehouseAvailableTablesModel } from '../../models/WarehouseAvailableTablesModel/WarehouseAvailableTablesModel';
 import { projectAdapterFromConfig } from '../../projectAdapters/projectAdapter';
 import {
-    CompiledQuery,
     applyLimitToSqlQuery,
     buildQuery,
+    CompiledQuery,
 } from '../../queryBuilder';
 import { compileMetricQuery } from '../../queryCompiler';
 import { SchedulerClient } from '../../scheduler/SchedulerClient';
@@ -201,10 +203,10 @@ import {
     validatePagination,
 } from './resultsPagination';
 import {
-    type PaginateDashboardChartArgs,
-    type PaginateMetricQueryArgs,
-    type PaginateQueryUuidArgs,
-    type PaginateSavedChartArgs,
+    type ExecuteAsyncDashboardChartQueryArgs,
+    type ExecuteAsyncMetricQueryArgs,
+    type ExecuteAsyncSavedChartQueryArgs,
+    type GetAsyncQueryResultsArgs,
 } from './types';
 
 type ProjectServiceArguments = {
@@ -1778,13 +1780,13 @@ export class ProjectService extends BaseService {
         });
     }
 
-    async runPaginatedQueryUuid({
+    async getAsyncQueryResults({
         user,
         projectUuid,
         queryUuid,
         page = 1,
         pageSize,
-    }: PaginateQueryUuidArgs): Promise<ApiPaginatedQueryResults> {
+    }: GetAsyncQueryResultsArgs): Promise<ApiGetAsyncQueryResults> {
         if (!isUserWithOrg(user)) {
             throw new ForbiddenError('User is not part of an organization');
         }
@@ -1806,13 +1808,37 @@ export class ProjectService extends BaseService {
             throw new ForbiddenError();
         }
 
-        const { metricQuery, context } = queryHistory;
+        const { metricQuery, context, status } = queryHistory;
+
+        switch (status) {
+            case QueryHistoryStatus.CANCELLED:
+            case QueryHistoryStatus.PENDING:
+                return {
+                    status,
+                    queryUuid,
+                    metricQuery,
+                    fields: queryHistory.fields,
+                };
+            case QueryHistoryStatus.ERROR:
+                return {
+                    status,
+                    queryUuid,
+                    metricQuery,
+                    fields: queryHistory.fields,
+                    error: queryHistory.error,
+                };
+            case QueryHistoryStatus.READY:
+                break;
+            default:
+                return assertUnreachable(status, 'Unknown query status');
+        }
 
         const explore = await this.getExplore(
             user,
             projectUuid,
             metricQuery.exploreName,
         );
+
         const { warehouseClient, sshTunnel } = await this._getWarehouseClient(
             projectUuid,
             await this.getWarehouseCredentials(projectUuid, user.userUuid),
@@ -1833,7 +1859,10 @@ export class ProjectService extends BaseService {
             query_context: context,
         };
 
-        const defaultedPageSize = pageSize ?? queryHistory.defaultPageSize;
+        const defaultedPageSize =
+            pageSize ??
+            queryHistory.defaultPageSize ??
+            DEFAULT_RESULTS_PAGE_SIZE;
 
         validatePagination({
             pageSize: defaultedPageSize,
@@ -1843,74 +1872,111 @@ export class ProjectService extends BaseService {
 
         const { result, durationMs } = await measureTime(
             () =>
-                warehouseClient.getPaginatedResults(
-                    {
-                        page,
-                        pageSize: defaultedPageSize,
-                        tags: queryTags,
-                        ...(queryHistory.warehouseQueryId
-                            ? {
-                                  queryId: queryHistory.warehouseQueryId,
-                                  queryMetadata:
-                                      queryHistory.warehouseQueryMetadata,
-                              }
-                            : { sql: queryHistory.compiledSql }),
-                    },
-                    formatter,
-                ),
-            'getPaginatedResults',
+                warehouseClient
+                    .getAsyncQueryResults(
+                        {
+                            page,
+                            pageSize: defaultedPageSize,
+                            tags: queryTags,
+                            queryId: queryHistory.warehouseQueryId,
+                            sql: queryHistory.compiledSql,
+                            queryMetadata: queryHistory.warehouseQueryMetadata,
+                        },
+                        formatter,
+                    )
+                    .catch((e) => ({ errorMessage: getErrorMessage(e) })),
+            'getAsyncQueryResults',
             this.logger,
             context,
         );
 
-        await sshTunnel.disconnect();
+        if ('errorMessage' in result) {
+            await this.queryHistoryModel.update(
+                queryHistory.queryUuid,
+                projectUuid,
+                user.userUuid,
+                {
+                    status: QueryHistoryStatus.ERROR,
+                    error: result.errorMessage,
+                },
+            );
 
-        const {
-            rows,
-            pageCount: totalPageCount,
-            totalRows: totalResults,
-        } = result;
+            return {
+                status: QueryHistoryStatus.ERROR,
+                error: result.errorMessage,
+                queryUuid: queryHistory.queryUuid,
+                metricQuery,
+                fields: queryHistory.fields,
+            };
+        }
+
+        const roundedDurationMs = Math.round(durationMs);
+
+        /**
+         * Update the query history with non null values
+         * defaultPageSize is null when user never fetched the results - we don't send pagination params to the query execution endpoint
+         * warehouseExecutionTimeMs is null when warehouse doesn't support async queries - query is only executed when user fetches results
+         * totalRowCount is null when warehouse doesn't support async queries - query is only executed when user fetches results
+         */
+        if (
+            queryHistory.defaultPageSize === null ||
+            queryHistory.warehouseExecutionTimeMs === null ||
+            queryHistory.totalRowCount === null
+        ) {
+            await this.queryHistoryModel.update(
+                queryHistory.queryUuid,
+                projectUuid,
+                user.userUuid,
+                {
+                    ...(queryHistory.defaultPageSize === null
+                        ? { default_page_size: defaultedPageSize }
+                        : {}),
+                    ...(queryHistory.warehouseExecutionTimeMs === null
+                        ? {
+                              warehouse_execution_time_ms: roundedDurationMs,
+                          }
+                        : {}),
+                    ...(queryHistory.totalRowCount === null
+                        ? { total_row_count: result.totalRows }
+                        : {}),
+                },
+            );
+        }
+
+        await sshTunnel.disconnect();
 
         const { nextPage, previousPage } = getNextAndPreviousPage(
             page,
-            totalPageCount,
+            result.pageCount,
         );
 
         return {
-            rows,
-            totalPageCount,
-            totalResults,
+            rows: result.rows,
+            totalPageCount: result.pageCount,
+            totalResults: result.totalRows,
             queryUuid: queryHistory.queryUuid,
             fields: queryHistory.fields,
             metricQuery,
-            pageSize: rows.length,
+            pageSize: result.rows.length,
             page,
             nextPage,
             previousPage,
-            initialQueryExecutionMs: queryHistory.warehouseExecutionTimeMs,
-            resultsPageExecutionMs: Math.round(durationMs),
+            initialQueryExecutionMs:
+                queryHistory.warehouseExecutionTimeMs ?? roundedDurationMs,
+            resultsPageExecutionMs: roundedDurationMs,
+            status,
         };
     }
 
-    private async runPaginatedQuery<
-        TFormattedRow extends Record<string, unknown>,
-    >(
-        args: PaginateMetricQueryArgs & {
-            user: SessionUser;
-            projectUuid: string;
-            invalidateCache?: boolean;
+    private async executeAsyncQuery(
+        args: ExecuteAsyncMetricQueryArgs & {
+            queryTags: RunQueryTags;
             exploreName: string;
-            granularity?: DateGranularity;
-            queryTags: RunQueryTags; // We already have context in the context parameter
         },
-        requestParameters: PaginatedQueryRequestParams,
-        rowFormatter?: (
-            row: Record<string, unknown>,
-            fields: ItemsMap,
-        ) => TFormattedRow,
+        requestParameters: ExecuteAsyncQueryRequestParams,
     ) {
         return wrapSentryTransaction(
-            'ProjectService.runPaginatedQuery',
+            'ProjectService.executeAsyncQuery',
             {},
             async (span) => {
                 const {
@@ -1918,8 +1984,6 @@ export class ProjectService extends BaseService {
                     projectUuid,
                     context,
                     granularity,
-                    page = 1,
-                    pageSize = DEFAULT_RESULTS_PAGE_SIZE,
                     queryTags,
                     exploreName,
                 } = args;
@@ -1946,12 +2010,6 @@ export class ProjectService extends BaseService {
                         throw new ForbiddenError();
                     }
 
-                    validatePagination({
-                        pageSize,
-                        page,
-                        queryMaxLimit: this.lightdashConfig.query.maxPageSize,
-                    });
-
                     const explore = await this.getExplore(
                         user,
                         projectUuid,
@@ -1976,8 +2034,6 @@ export class ProjectService extends BaseService {
                         'warehouse.type',
                         warehouseClient.credentials.type,
                     );
-                    span.setAttribute('query.page', page);
-                    span.setAttribute('query.pageSize', pageSize);
 
                     const { metricQuery } = args;
                     const userAttributes =
@@ -2043,74 +2099,63 @@ export class ProjectService extends BaseService {
                         );
                     }
 
-                    const formatter = (
-                        row: Record<string, unknown>,
-                    ): TFormattedRow =>
-                        rowFormatter
-                            ? rowFormatter(row, fieldsMap)
-                            : (row as TFormattedRow);
+                    const { queryUuid: queryHistoryUuid } =
+                        await this.queryHistoryModel.create({
+                            projectUuid,
+                            organizationUuid,
+                            createdByUserUuid: user.userUuid,
+                            context,
+                            fields: fieldsMap,
+                            compiledSql: query,
+                            requestParameters,
+                            metricQuery,
+                        });
 
-                    const { result, durationMs } = await measureTime(
-                        () =>
-                            warehouseClient.getPaginatedResults(
+                    // Trigger query in the background, update query history when complete
+                    warehouseClient
+                        .executeAsyncQuery({
+                            sql: query,
+                            tags: queryTags,
+                        })
+                        .then(
+                            ({
+                                queryId,
+                                queryMetadata,
+                                totalRows,
+                                durationMs,
+                            }) =>
+                                this.queryHistoryModel.update(
+                                    queryHistoryUuid,
+                                    projectUuid,
+                                    user.userUuid,
+                                    {
+                                        warehouse_query_id: queryId,
+                                        warehouse_query_metadata: queryMetadata,
+                                        status: QueryHistoryStatus.READY,
+                                        error: null,
+                                        warehouse_execution_time_ms:
+                                            durationMs !== null
+                                                ? Math.round(durationMs)
+                                                : null,
+                                        total_row_count: totalRows,
+                                    },
+                                ),
+                        )
+                        .catch((e) =>
+                            this.queryHistoryModel.update(
+                                queryHistoryUuid,
+                                projectUuid,
+                                user.userUuid,
                                 {
-                                    sql: query,
-                                    page,
-                                    pageSize,
-                                    tags: queryTags,
+                                    status: QueryHistoryStatus.ERROR,
+                                    error: getErrorMessage(e),
                                 },
-                                formatter,
                             ),
-                        'getPaginatedResults',
-                        this.logger,
-                        context,
-                    );
-
-                    const {
-                        rows,
-                        pageCount: totalPageCount,
-                        totalRows: totalResults,
-                        queryId,
-                    } = result;
-
-                    await sshTunnel.disconnect();
-
-                    const { nextPage, previousPage } = getNextAndPreviousPage(
-                        page,
-                        totalPageCount,
-                    );
-
-                    const roundedDurationMs = Math.round(durationMs);
-                    const queryHistory = await this.queryHistoryModel.create({
-                        projectUuid,
-                        organizationUuid,
-                        createdByUserUuid: user.userUuid,
-                        defaultPageSize: pageSize,
-                        context,
-                        fields: fieldsMap,
-                        compiledSql: query,
-                        requestParameters,
-                        metricQuery,
-                        totalRowCount: totalResults,
-                        warehouseQueryId: queryId,
-                        warehouseExecutionTimeMs: roundedDurationMs,
-                        warehouseQueryMetadata: result.warehouseQueryMetadata,
-                    });
+                        )
+                        .finally(() => sshTunnel.disconnect());
 
                     return {
-                        rows,
-                        queryUuid: queryHistory.queryUuid,
-                        totalPageCount,
-                        totalResults,
-                        page,
-                        pageSize: rows.length,
-                        nextPage,
-                        previousPage,
-                        fields: fieldsMap,
-                        metricQuery,
-                        // Since this is the initial query execution, we use the same value for both initial and query execution seconds
-                        initialQueryExecutionMs: roundedDurationMs,
-                        resultsPageExecutionMs: roundedDurationMs,
+                        queryUuid: queryHistoryUuid,
                     };
                 } catch (e) {
                     span.setStatus({
@@ -2125,15 +2170,13 @@ export class ProjectService extends BaseService {
         );
     }
 
-    async runPaginatedMetricQuery({
+    async executeAsyncMetricQuery({
         user,
         projectUuid,
-        dateZoomGranularity,
+        granularity,
         context,
         metricQuery,
-        page,
-        pageSize,
-    }: PaginateMetricQueryArgs): Promise<ApiPaginatedQueryResults> {
+    }: ExecuteAsyncMetricQueryArgs): Promise<ApiExecuteAsyncQueryResults> {
         if (!isUserWithOrg(user)) {
             throw new ForbiddenError('User is not part of an organization');
         }
@@ -2161,10 +2204,8 @@ export class ProjectService extends BaseService {
             );
         }
 
-        const requestParameters: PaginatedMetricQueryRequestParams = {
+        const requestParameters: ExecuteAsyncMetricQueryRequestParams = {
             context,
-            page,
-            pageSize,
             query: metricQuery,
         };
 
@@ -2176,7 +2217,7 @@ export class ProjectService extends BaseService {
             query_context: context,
         };
 
-        return this.runPaginatedQuery(
+        const { queryUuid } = await this.executeAsyncQuery(
             {
                 user,
                 metricQuery,
@@ -2184,24 +2225,24 @@ export class ProjectService extends BaseService {
                 exploreName: metricQuery.exploreName,
                 context,
                 queryTags,
-                granularity: dateZoomGranularity,
-                page,
-                pageSize,
+                granularity,
             },
             requestParameters,
-            formatRow,
         );
+
+        return {
+            queryUuid,
+            appliedDashboardFilters: null,
+        };
     }
 
-    async runPaginatedSavedChartQuery({
+    async executeAsyncSavedChartQuery({
         user,
         projectUuid,
         chartUuid,
         versionUuid,
-        page,
-        pageSize,
         context,
-    }: PaginateSavedChartArgs): Promise<ApiPaginatedQueryResults> {
+    }: ExecuteAsyncSavedChartQueryArgs): Promise<ApiExecuteAsyncQueryResults> {
         // Check user is in organization
         if (!isUserWithOrg(user)) {
             throw new ForbiddenError('User does not belong to an organization');
@@ -2256,10 +2297,8 @@ export class ProjectService extends BaseService {
             user.userUuid,
         );
 
-        const requestParameters: PaginatedSavedChartRequestParams = {
+        const requestParameters: ExecuteAsyncSavedChartRequestParams = {
             context,
-            page,
-            pageSize,
             chartUuid,
             versionUuid,
         };
@@ -2273,24 +2312,26 @@ export class ProjectService extends BaseService {
             query_context: context,
         };
 
-        return this.runPaginatedQuery(
+        const { queryUuid } = await this.executeAsyncQuery(
             {
                 user,
                 projectUuid,
                 exploreName: savedChartTableName,
-                page,
-                pageSize,
                 context,
                 queryTags,
                 invalidateCache: false,
                 metricQuery,
             },
             requestParameters,
-            formatRow,
         );
+
+        return {
+            queryUuid,
+            appliedDashboardFilters: null,
+        };
     }
 
-    async runPaginatedDashboardChartQuery({
+    async executeAsyncDashboardChartQuery({
         user,
         projectUuid,
         chartUuid,
@@ -2298,10 +2339,8 @@ export class ProjectService extends BaseService {
         dashboardFilters,
         dashboardSorts,
         granularity,
-        page,
-        pageSize,
         context,
-    }: PaginateDashboardChartArgs): Promise<ApiPaginatedQueryResults> {
+    }: ExecuteAsyncDashboardChartQueryArgs): Promise<ApiExecuteAsyncQueryResults> {
         if (!isUserWithOrg(user)) {
             throw new ForbiddenError('User is not part of an organization');
         }
@@ -2356,7 +2395,7 @@ export class ProjectService extends BaseService {
         );
 
         const tables = Object.keys(explore.tables);
-        const appliedDashboardFilters = {
+        const appliedDashboardFilters: DashboardFilters = {
             dimensions: getDashboardFilterRulesForTables(
                 tables,
                 dashboardFilters.dimensions,
@@ -2403,10 +2442,8 @@ export class ProjectService extends BaseService {
             };
         }
 
-        const requestParameters: PaginatedDashboardChartRequestParams = {
+        const requestParameters: ExecuteAsyncDashboardChartRequestParams = {
             context,
-            page,
-            pageSize,
             chartUuid,
             dashboardUuid,
             dashboardFilters,
@@ -2424,7 +2461,7 @@ export class ProjectService extends BaseService {
             query_context: context,
         };
 
-        return this.runPaginatedQuery(
+        const { queryUuid } = await this.executeAsyncQuery(
             {
                 user,
                 projectUuid,
@@ -2433,13 +2470,15 @@ export class ProjectService extends BaseService {
                 context,
                 queryTags,
                 invalidateCache: false,
-                page,
-                pageSize,
                 granularity,
             },
             requestParameters,
-            formatRow,
         );
+
+        return {
+            queryUuid,
+            appliedDashboardFilters,
+        };
     }
 
     private async runQueryAndFormatRows({
@@ -3024,6 +3063,7 @@ export class ProjectService extends BaseService {
                             ).filter((metricOverrideKey) =>
                                 metricQuery.metrics.includes(metricOverrideKey),
                             ).length,
+                            limit: metricQueryWithLimit.limit,
                         },
                     });
                     this.logger.debug(