+++ packages/common/src/compiler/exploreCompiler.mock.ts
@@ -38,13 +38,18 @@ export const warehouseClientMock: WarehouseClient = {
         });
         return Promise.resolve();
     },
-    getPaginatedResults: async () => ({
+    executeAsyncQuery: async () => ({
+        queryId: null,
+        queryMetadata: null,
+        totalRows: null,
+        durationMs: null,
+    }),
+    getAsyncQueryResults: async () => ({
         rows: [],
         fields: {},
         queryId: null,
         pageCount: 0,
         totalRows: 0,
-        warehouseQueryMetadata: null,
     }),
     runQuery: () =>
         Promise.resolve({