+++ packages/common/src/utils/convertCustomDimensionsToYaml.ts
@@ -97,7 +97,10 @@ const warehouseClientMock: WarehouseClient = {
     streamQuery() {
         throw new NotImplementedError('getCatalog not implemented');
     },
-    getPaginatedResults() {
+    executeAsyncQuery() {
+        throw new NotImplementedError('getCatalog not implemented');
+    },
+    getAsyncQueryResults() {
         throw new NotImplementedError('getCatalog not implemented');
     },
     test() {