+++ packages/warehouses/src/warehouseClients/SnowflakeWarehouseClient.test.ts
@@ -1,7 +1,7 @@
 import {
     DimensionType,
     type ResultRow,
-    type WarehousePaginatedResults,
+    type WarehouseGetAsyncQueryResults,
 } from '@lightdash/common';
 import { createConnection } from 'snowflake-sdk';
 import { Readable } from 'stream';
@@ -55,6 +55,8 @@ jest.mock('snowflake-sdk', () => ({
         execute: executeMock,
         destroy: jest.fn((callback) => callback(null, {})),
         getResultsFromQueryId: getResultsFromQueryIdMock,
+        getQueryStatus: jest.fn(() => 'SUCCESS'),
+        isStillRunning: jest.fn(() => false),
     })),
 }));
 
@@ -83,29 +85,23 @@ describe('SnowflakeWarehouseClient', () => {
             expectedWarehouseSchema,
         );
     });
-    describe('getPaginatedResults', () => {
+    describe('getAsyncQueryResults', () => {
         it('should return raw results', async () => {
             const client = new SnowflakeWarehouseClient(credentials);
 
             // Execute test
-            const result = await client.getPaginatedResults({
+            const result = await client.getAsyncQueryResults({
                 sql: 'SELECT * FROM test',
+                queryId: 'queryId',
                 page: 1,
                 pageSize: 10,
                 tags: {},
                 timezone: 'UTC',
                 values: [],
+                queryMetadata: null,
             });
 
             // Assertions
-
-            expect(executeMock).toHaveBeenLastCalledWith({
-                sqlText: 'SELECT * FROM test',
-                asyncExec: true,
-                binds: undefined,
-                complete: expect.any(Function),
-            });
-
             expect(getResultsFromQueryIdMock).toHaveBeenCalledWith({
                 sqlText: '',
                 queryId: 'queryId',
@@ -117,8 +113,7 @@ describe('SnowflakeWarehouseClient', () => {
                 queryId: 'queryId',
                 pageCount: 1,
                 totalRows: 1,
-                warehouseQueryMetadata: null,
-            } satisfies WarehousePaginatedResults<Record<string, unknown>>);
+            } satisfies WarehouseGetAsyncQueryResults<Record<string, unknown>>);
         });
 
         it('should return formatted results when using a formatter', async () => {
@@ -138,7 +133,7 @@ describe('SnowflakeWarehouseClient', () => {
                 );
             }
 
-            const result = await client.getPaginatedResults(
+            const result = await client.getAsyncQueryResults(
                 {
                     sql: 'SELECT * FROM test',
                     page: 1,
@@ -146,6 +141,8 @@ describe('SnowflakeWarehouseClient', () => {
                     tags: {},
                     timezone: 'UTC',
                     values: [],
+                    queryId: 'queryId',
+                    queryMetadata: null,
                 },
                 formatter,
             );
@@ -153,13 +150,6 @@ describe('SnowflakeWarehouseClient', () => {
             const expectedFormattedRow = formatter(expectedRow);
 
             // Assertions
-            expect(executeMock).toHaveBeenLastCalledWith({
-                sqlText: 'SELECT * FROM test',
-                asyncExec: true,
-                binds: undefined,
-                complete: expect.any(Function),
-            });
-
             expect(getResultsFromQueryIdMock).toHaveBeenCalledWith({
                 sqlText: '',
                 queryId: 'queryId',
@@ -171,14 +161,13 @@ describe('SnowflakeWarehouseClient', () => {
                 queryId: 'queryId',
                 pageCount: 1,
                 totalRows: 1,
-                warehouseQueryMetadata: null,
-            } satisfies WarehousePaginatedResults<ResultRow>);
+            } satisfies WarehouseGetAsyncQueryResults<ResultRow>);
         });
 
         it('should not execute any query when using queryId', async () => {
             const client = new SnowflakeWarehouseClient(credentials);
 
-            const result = await client.getPaginatedResults({
+            const result = await client.getAsyncQueryResults({
                 queryId: 'queryId',
                 queryMetadata: null,
                 page: 1,
@@ -186,6 +175,7 @@ describe('SnowflakeWarehouseClient', () => {
                 tags: {},
                 timezone: 'UTC',
                 values: [],
+                sql: 'SELECT * FROM test',
             });
 
             // Assertions
@@ -208,8 +198,7 @@ describe('SnowflakeWarehouseClient', () => {
                 queryId: 'queryId',
                 pageCount: 1,
                 totalRows: 1,
-                warehouseQueryMetadata: null,
-            } satisfies WarehousePaginatedResults<Record<string, unknown>>);
+            } satisfies WarehouseGetAsyncQueryResults<Record<string, unknown>>);
         });
     });
 });